requirements.txt

google-cloud-aiplatform[agent_engines,adk]>=1.112.0
mcp>=1.9.0
httpx>=0.27.0
python-dotenv>=1.0.1
